sheep: fix 32-bit integer overflow in stat_sheep()

+++ sheep/store.c
@@ -101,8 +101,8 @@ static int stat_sheep(uint64_t *store_size, uint64_t *store_free, uint32_t epoch
 
 	closedir(dir);
 
-	*store_size = vs.f_frsize * vs.f_bfree + used;
-	*store_free = vs.f_frsize * vs.f_bfree;
+	*store_size = (uint64_t)vs.f_frsize * vs.f_bfree + used;
+	*store_free = (uint64_t)vs.f_frsize * vs.f_bfree;
 
 	return SD_RES_SUCCESS;
 }
